final_backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-multipart
python-dotenv
pytz>=2022.7
